requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
django==2.1.12
djangorestframework-jwt==1.11.0
djangorestframework==3.10.3
idna-ssl==1.1.0 ; python_version < '3.7'
idna==2.8
multidict==4.5.2
psycopg2==2.7.3
pyjwt==1.7.1
pytz==2019.2
requests-cache==0.5.2
requests==2.22.0
typing-extensions==3.7.4 ; python_version < '3.7'
urllib3==1.25.3
yarl==1.3.0
django-filter==2.2.0